Communication/requirements.txt
PyYYAML==5.4.1
future==0.18.2
construct==2.10.67
configparser==5.0.2
google==3.0.0
google-api-core==1.31.0
google-api-python-client==2.14.0
google-cloud==0.34.0
google-cloud-biggquery==2.22.0
google-cloud-core==1.7.1
logger==1.4
pip==21.2.3
protobuf==3.14.0
pycparser==2.20
pyparsing==2.4.7
pyserial==3.5
pyzmq==22.2.1
numpy==1.17.4
packaging==21.0
python-dateutil==2.8.1
setuptools==57.4.0
Jinja2==2.10.1
typing-extensions==3.10.0.0
dotmap==1.3.17
smbus2~=0.3.0
scipy==1.3.3
matplotlib==3.2.0
toml==0.10.0
keyboard=0.13.5
